+++ src/DependencyInjection/DiscoToolbarExtension.php
@@ -18,10 +18,27 @@ declare(strict_types=1);
  *
  * ******************************************************************* */
 
-namespace MarcinOrlowski\DiscoToolbar\Dto;
+namespace MarcinOrlowski\DiscoToolbar\DependencyInjection;
 
-enum IconType: string
+use Symfony\Component\Config\FileLocator;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Extension\Extension;
+use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
+
+/**
+ * DiscoToolbar Extension
+ *
+ * Loads bundle services configuration
+ */
+class DiscoToolbarExtension extends Extension
 {
-    case FONT_AWESOME = 'fa';
-    case TEXT = 'text';
+    public function load(array $configs, ContainerBuilder $container): void
+    {
+        $loader = new YamlFileLoader(
+            $container,
+            new FileLocator(\dirname(__DIR__) . '/../Resources/config')
+        );
+
+        $loader->load('services.yaml');
+    }
 }