+++ CHANGES.md
@@ -1,9 +1,15 @@
+![img](img/banner.webp)
+
 # Changelog
 
-## 1.6.0 (TBD)
+## 2.0.0 (2026-01-30)
 
+- [#044] Renamed package to `marcin-orlowski/disco-toolbar-symfony`
+- [#047] Rebranded from DiscoDevBar to DiscoToolbar
+- [#041] Added Symfony 8 support
+- [#042] Added project logo!
 - [#035] Added customizable background colors for breathing stripes
-- [#033] Added close widget type to dismiss devbar
+- [#033] Added close widget type to dismiss toolbar
 
 ## 1.5.2 (2025-12-12)